+++ src/video_codecs/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "amf")]
+pub mod amf;
 pub mod internal;
 #[cfg(any(target_os = "macos", target_os = "ios"))]
 pub mod internal_hwa;